Add prettier
We want to use Prettier to automate stylistic rules for our codebases.
Prettier is an established, opinionated, tool in the community and is
used by many companies and open-source codebases.

We believe that adding Prettier to our shared linting config will help
further align Pg with the JavaScript community as a whole.